+++ packages/pluggableWidgets/combobox-web/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- We fixed an issue with filter input not being visible in some cases.
+
+- We fixed an issue with Backspace key clearing input even though Clearable setting was set to false.
+
 ## [2.7.1] - 2026-01-26
 
 ### Fixed